feat[#7] : 카테고리가 삭제될 때 shopcategory 도 삭제되도록 구현
- 이때 root category의 경우 shop까지 같이 삭제되도록 구현

+++ backend/src/main/java/com/handong/rebon/shop/domain/category/ShopCategory.java
@@ -29,8 +29,10 @@ public class ShopCategory extends BaseEntity {
         this.category = category;
     }
 
-    public void delete() {
+    public void delete(boolean isRootCategory) {
         this.deleteContent();
-        this.shop.delete();
+        if (isRootCategory) {
+            this.shop.delete();
+        }
     }
 }
